Projeto v2.0

+++ Course/Entities/OrderItem.cs
@@ -16,9 +16,10 @@
             Price = price;
             Product = product;
         }
+
         public double SubTotal()
         {
-            return Quantity * Price;
+            return (double) Quantity * Price;
         }
     }
 }
